chore(chromes): rename AdminChromeLayout back to AdminSSChrome
